feat: Update meal details screen to show favorite status dynamically
This commit updates the meal details screen to dynamically show the favorite status of the meal. The favoriteMealsProvider is now watched to determine if the meal is in the list of favorite meals. If it is, the star icon is filled, indicating that the meal is a favorite. This change improves the user experience by providing real-time feedback on the favorite status of the meal.